services/account/requirements.txt
black==22.3.0
Flask==2.1.2
google-cloud-pubsub==2.12.0
huey==2.4.3
pg8000==1.26.1
pylint==2.13.9
requests==2.27.1
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
